+++ slick2d-core/src/main/java/org/newdawn/slick/Music.java
@@ -47,7 +47,7 @@ public class Music {
 	/** True if the music is playing */
 	private boolean playing;
 	/** The list of listeners waiting for notification that the music ended */
-	private ArrayList listeners = new ArrayList();
+	private ArrayList<MusicListener> listeners = new ArrayList<>();
 	/** The volume of this music */
 	private float volume = 1.0f;
 	/** Start gain for fading in/out */
@@ -288,6 +288,7 @@ public class Music {
 	 * Stop the music playing
 	 */
 	public void stop() {
+	    playing = false;
 		sound.stop();
 	}
 	
@@ -364,7 +365,7 @@ public class Music {
        
 		if (fadeTime > 0) {
 			fadeTime -= delta;
-			if (fadeTime < 0) {
+			if (fadeTime <= 0) {
 				fadeTime = 0;
 				if (stopAfterFade) {
 					stop();