+++ slick2d-core/src/test/java/org/newdawn/slick/tests/util/TestWithLWJGL.java
@@ -1,0 +1,15 @@
+package org.newdawn.slick.tests.util;
+
+import java.io.File;
+
+import org.testng.annotations.BeforeTest;
+
+public class TestWithLWJGL {
+
+    @BeforeTest
+    public void setLWJGLLibraryPath() {
+        System.setProperty("org.lwjgl.librarypath",
+            new File(".", "target" + File.separator + "natives").getAbsolutePath());
+    }
+
+}