+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Fixed `RoundedRectangle.contains()` after calling `setWidth()` / `setHeight()` thanks to [Joshua Hertlein](https://bitbucket.org/kevglass/slick/issue/41/roundedrectangles-setwidth-and-setheight)
 * Fixed `AppContainer.setMouseCursor()` by using correct image dimensions thanks to [Peter W](https://bitbucket.org/kevglass/slick/issue/40/using-wrong-variable)
 * Fixed `SpriteSheetFont.drawString()` with indexes thanks to [Emmanuel Rousseau](https://bitbucket.org/kevglass/slick/issue/39/rendering-bug-with-spritesheetfont)
+* Fixed `Music` [not stopping](https://bitbucket.org/kevglass/slick/issue/30/musicfade-stopafterfade-does-not-work) when using `fade(.., ..., true)`
 
 ## v1.0.0 (2014-04-16)